begin adding list table, adjust models, views and unit tests to accomodate
lists table supports the requirement to support multiple lists.  That requires
the creation of a new django model and consequently requires new migrations.
The list table is a foreign key in the items table and therefore the
Items table must be updated with that info (and assoc. migrations). The tests
are updated to check for the list assocaited with a new item and views must
also be updated to add the list field to a new Item row/object.  Func tests
are still failing as before (line 86 in test_start_and_retrieve_list of
functest/tests.py